requirements.txt
pandas>=2.0
yfinance>=0.2
pyyaml>=6.0
tqdm>=4.0
numpy>=1.22
python-dateutil>=2.8